fix tests: end dependence of registry in tests
