+++ cmd/limactl/prune.go
@@ -5,6 +5,7 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"maps"
 	"os"
 
@@ -12,6 +13,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/lima-vm/lima/v2/pkg/downloader"
+	"github.com/lima-vm/lima/v2/pkg/driverutil"
 	"github.com/lima-vm/lima/v2/pkg/limatype"
 	"github.com/lima-vm/lima/v2/pkg/limayaml"
 	"github.com/lima-vm/lima/v2/pkg/store"
@@ -94,6 +96,9 @@ func knownLocations(ctx context.Context) (map[string]limatype.File, error) {
 		y, err := limayaml.Load(ctx, b, t.Name)
 		if err != nil {
 			return nil, err
+		}
+		if err := driverutil.ResolveVMType(y, t.Name); err != nil {
+			return nil, fmt.Errorf("failed to accept config for %q: %w", t.Name, err)
 		}
 		maps.Copy(locations, locationsFromLimaYAML(y))
 	}
