+++ cmd/limactl/template.go
@@ -13,9 +13,10 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
+	"github.com/lima-vm/lima/v2/pkg/driverutil"
 	"github.com/lima-vm/lima/v2/pkg/limatmpl"
+	"github.com/lima-vm/lima/v2/pkg/limatype/dirnames"
 	"github.com/lima-vm/lima/v2/pkg/limayaml"
-	"github.com/lima-vm/lima/v2/pkg/store/dirnames"
 	"github.com/lima-vm/lima/v2/pkg/yqutil"
 )
 
@@ -86,6 +87,9 @@ func fillDefaults(ctx context.Context, tmpl *limatmpl.Template) error {
 	tmpl.Config, err = limayaml.Load(ctx, tmpl.Bytes, filePath)
 	if err == nil {
 		tmpl.Bytes, err = limayaml.Marshal(tmpl.Config, false)
+	}
+	if err := driverutil.ResolveVMType(tmpl.Config, filePath); err != nil {
+		return fmt.Errorf("failed to accept config for %q: %w", filePath, err)
 	}
 	return err
 }
@@ -245,6 +249,9 @@ func templateValidateAction(cmd *cobra.Command, args []string) error {
 		y, err := limayaml.Load(ctx, tmpl.Bytes, filePath)
 		if err != nil {
 			return err
+		}
+		if err := driverutil.ResolveVMType(y, filePath); err != nil {
+			return fmt.Errorf("failed to accept config for %q: %w", filePath, err)
 		}
 		if err := limayaml.Validate(y, false); err != nil {
 			return fmt.Errorf("failed to validate YAML file %q: %w", arg, err)
