+++ pkg/limayaml/defaults.go
@@ -30,14 +30,13 @@ import (
 	"github.com/lima-vm/lima/v2/pkg/instance/hostname"
 	"github.com/lima-vm/lima/v2/pkg/ioutilx"
 	"github.com/lima-vm/lima/v2/pkg/limatype"
+	"github.com/lima-vm/lima/v2/pkg/limatype/dirnames"
+	"github.com/lima-vm/lima/v2/pkg/limatype/filenames"
 	"github.com/lima-vm/lima/v2/pkg/localpathutil"
 	. "github.com/lima-vm/lima/v2/pkg/must"
 	"github.com/lima-vm/lima/v2/pkg/networks"
 	"github.com/lima-vm/lima/v2/pkg/osutil"
 	"github.com/lima-vm/lima/v2/pkg/ptr"
-	"github.com/lima-vm/lima/v2/pkg/registry"
-	"github.com/lima-vm/lima/v2/pkg/store/dirnames"
-	"github.com/lima-vm/lima/v2/pkg/store/filenames"
 	"github.com/lima-vm/lima/v2/pkg/version"
 )
 
@@ -210,14 +209,14 @@ func FillDefault(ctx context.Context, y, d, o *limatype.LimaYAML, filePath strin
 	if o.VMType != nil {
 		y.VMType = o.VMType
 	}
-	if y.VMType != nil && *y.VMType != "" && *y.VMType != "default" {
-		logrus.Debugf("ResolveVMType: VMType %q is explicitly specified in %q", *y.VMType, filePath)
-		_, _, exists := registry.Get(*y.VMType)
-		if !exists {
-			logrus.Warnf("ResolveVMType: VMType %q is not registered", *y.VMType)
-		}
-		*y.VMType = limatype.NewVMType(*y.VMType)
-	}
+	// if y.VMType != nil && *y.VMType != "" && *y.VMType != "default" {
+	// 	logrus.Debugf("ResolveVMType: VMType %q is explicitly specified in %q", *y.VMType, filePath)
+	// 	_, _, exists := registry.Get(*y.VMType)
+	// 	if !exists {
+	// 		logrus.Warnf("ResolveVMType: VMType %q is not registered", *y.VMType)
+	// 	}
+	// 	*y.VMType = limatype.NewVMType(*y.VMType)
+	// }
 
 	if y.OS == nil {
 		y.OS = d.OS
